+++ cmd/tfproviderlintx/tfproviderlintx.go
@@ -1,4 +1,4 @@
-// The tfproviderlint command is a static checker for Terraform Providers.
+// The tfproviderlintx command is a static checker for Terraform Providers that includes extra checks.
 //
 // Each analyzer flag name is preceded by the analyzer name: -NAME.flag.
 // In addition, the -NAME flag itself controls whether the
@@ -9,11 +9,16 @@ package main
 import (
 	"github.com/bflad/tfproviderlint/helper/cmdflags"
 	"github.com/bflad/tfproviderlint/passes"
+	"github.com/bflad/tfproviderlint/xpasses"
+	"golang.org/x/tools/go/analysis"
 	"golang.org/x/tools/go/analysis/multichecker"
 )
 
 func main() {
 	cmdflags.AddVersionFlag()
 
-	multichecker.Main(passes.AllChecks...)
+	var analyzers []*analysis.Analyzer
+	analyzers = append(analyzers, passes.AllChecks...)
+	analyzers = append(analyzers, xpasses.AllChecks...)
+	multichecker.Main(analyzers...)
 }