+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 FEATURES
 
+* **New Command:** `tfproviderlintx`: Runs all standard and extra checks
 * **New Extra Check:** `XS001`: check for `map[string]*Schema` that `Description` is configured
 
 # v0.7.0